Add percentage and total test time to EventPro
Added a counter called total_test_time to the Profiler class
that gets incremented with each example time. This attribute
helps the profiler class's caller to know the total test time
of all runs. Some other changes are related with the percentage
took by the examples. Those changes use the total_test_time to
calculate that percentage and the EventProfFormatter and
RSpecListener methods that involved the reporting were updated.