Fix first order state leakages
Most of the state leakage cases are caused by examples or the code under
test mutating models that are shared between examples when not
`reload`'ed or `refind`'ed.

Those leakages are hard to track down to the root cause and lead to
unpleasant surprises that usually accumulate error. When the root cause
changes its behavior, the problem becomes apparent with numerous
failures.

Co-authored-by: Jaimerson Araújo <jaimersonaraujo@gmail.com>